requirements.txt
Pillow==3.3.1
CairoSVG==1.0.22
requests==2.9.1
